+++ src/scenes/game_scene.cpp
@@ -539,9 +539,9 @@ void GameScene::set_space_texs(SpaceType type,const Texture* ceiling,const Textu
   const int space_id = SpaceTypes::value_of(type);
 
   // Ceiling & Floor textures are flipped due to using opposite values in Dantares ctor.
-  if(ceiling != nullptr) { dantares_->SetFloorTexture(space_id,static_cast<int>(ceiling->handle())); }
-  if(wall != nullptr) { dantares_->SetWallTexture(space_id,static_cast<int>(wall->handle())); }
-  if(floor != nullptr) { dantares_->SetCeilingTexture(space_id,static_cast<int>(floor->handle())); }
+  if(ceiling != nullptr) { dantares_->SetFloorTexture(space_id,ceiling->handle()); }
+  if(wall != nullptr) { dantares_->SetWallTexture(space_id,wall->handle()); }
+  if(floor != nullptr) { dantares_->SetCeilingTexture(space_id,floor->handle()); }
 }
 
 } // namespace ekoscape
